Refract RepListCommand to DSLCommand
Tip from @Nightdavisao

https://github.com/LorittaBot/Loritta/pull/2137#issuecomment-630992115
