+++ loritta-plugins/helping-hands/src/main/kotlin/net/perfectdreams/loritta/plugin/helpinghands/HelpingHandsPlugin.kt
@@ -6,6 +6,7 @@ import net.perfectdreams.loritta.platform.discord.LorittaDiscord
 import net.perfectdreams.loritta.platform.discord.plugin.LorittaDiscordPlugin
 import net.perfectdreams.loritta.plugin.helpinghands.commands.CoinFlipBetCommand
 import net.perfectdreams.loritta.plugin.helpinghands.commands.DailyInactivityTaxExecutor
+import net.perfectdreams.loritta.plugin.helpinghands.commands.RepListCommand
 import net.perfectdreams.loritta.plugin.helpinghands.utils.DailyInactivityTaxUtils
 
 class HelpingHandsPlugin(name: String, loritta: LorittaBot) : LorittaDiscordPlugin(name, loritta) {
@@ -13,6 +14,7 @@ class HelpingHandsPlugin(name: String, loritta: LorittaBot) : LorittaDiscordPlug
 		loritta as LorittaDiscord
 
 		registerCommand(CoinFlipBetCommand.command(this, loritta))
+		registerCommand(RepListCommand.command(this, loritta));
 
 		if (loritta.isMaster)
 			launch(DailyInactivityTaxUtils.createAutoInactivityTask())
